EntregaSemana8/app_server/requirements.txt
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
pika==1.3.1
psycopg2==2.9.5
PyJWT==2.6.0